test(lightweight): working tests with jsdom
