+++ apps/lightweight/test/app/dashboard/page.test.tsx
@@ -1,0 +1,117 @@
+import userEvent from "@testing-library/user-event";
+import { http, UserRejectedRequestError } from "viem";
+import { generatePrivateKey, privateKeyToAddress } from "viem/accounts";
+import { mainnet, base } from "viem/chains";
+import { describe, expect, vi } from "vitest";
+import { mock } from "wagmi";
+
+import { forkedTest, rpcUrls } from "../../config";
+import { render, screen, waitFor, waitForElementToBeRemoved } from "../../providers";
+
+import Page from "@/app/dashboard/page";
+import { createConfig } from "@/lib/config";
+
+describe("connect wallet flow", () => {
+  forkedTest("handles user rejection gracefully", async ({ client }) => {
+    const account = privateKeyToAddress(generatePrivateKey());
+    const wagmiConfig = createConfig({
+      chains: [mainnet, base],
+      transports: {
+        [mainnet.id]: http(client.transport.url),
+        [base.id]: http(rpcUrls[base.id]),
+      },
+      connectors: [
+        mock({
+          accounts: [account],
+          features: {
+            connectError: new UserRejectedRequestError(new Error("Failed to connect.")),
+          },
+        }),
+      ],
+    });
+
+    render(<Page />, { wagmiConfig });
+
+    await userEvent.click(screen.getByText("Connect Wallet"));
+    await userEvent.click(await screen.findByText("Mock Connector"));
+
+    await waitFor(() => screen.findByText("Requesting Connection"));
+    await waitForElementToBeRemoved(screen.getByText("Requesting Connection"));
+
+    expect(screen.getByText("Request Cancelled")).toBeInTheDocument();
+  });
+
+  forkedTest("shows user address once connected", async ({ client }) => {
+    const account = privateKeyToAddress(generatePrivateKey());
+    const wagmiConfig = createConfig({
+      chains: [mainnet, base],
+      transports: {
+        [mainnet.id]: http(client.transport.url),
+        [base.id]: http(rpcUrls[base.id]),
+      },
+      connectors: [mock({ accounts: [account] })],
+    });
+
+    render(<Page />, { wagmiConfig });
+
+    await userEvent.click(screen.getByText("Connect Wallet"));
+    await userEvent.click(await screen.findByText("Mock Connector"));
+
+    await waitFor(() => screen.findByText("Requesting Connection"));
+    await waitForElementToBeRemoved(screen.getByText("Requesting Connection"));
+
+    expect(screen.getByText(`${account.slice(0, 6)}...${account.slice(-4)}`)).toBeInTheDocument();
+  });
+});
+
+describe("switch chain flow", () => {
+  forkedTest("switches to base without wallet and opens main app", async ({ client }) => {
+    const account = privateKeyToAddress(generatePrivateKey());
+    const wagmiConfig = createConfig({
+      chains: [mainnet, base],
+      transports: {
+        [mainnet.id]: http(client.transport.url),
+        [base.id]: http(rpcUrls[base.id]),
+      },
+      connectors: [mock({ accounts: [account] })],
+    });
+
+    render(<Page />, {
+      wagmiConfig,
+      routes: [{ element: <div>Switched to Base successfully!</div>, path: "base/earn" }],
+    });
+
+    window.open = vi.fn();
+
+    await userEvent.click(screen.getByRole("combobox"));
+    await userEvent.click(screen.getByText("Base"));
+
+    expect(screen.getByText("Switched to Base successfully!")).toBeInTheDocument();
+    expect(window.open).toHaveBeenCalledWith("https://app.morpho.org/base/earn", "_blank", "noopener,noreferrer");
+  });
+
+  forkedTest("switches to base with wallet and opens main app", async ({ client }) => {
+    const account = privateKeyToAddress(generatePrivateKey());
+    const wagmiConfig = createConfig({
+      chains: [mainnet, base],
+      transports: {
+        [mainnet.id]: http(client.transport.url),
+        [base.id]: http(rpcUrls[base.id]),
+      },
+      connectors: [mock({ accounts: [account], features: { defaultConnected: true } })],
+    });
+
+    render(<Page />, {
+      wagmiConfig,
+      routes: [{ element: <div>Switched to Base successfully!</div>, path: "base/earn" }],
+    });
+
+    window.open = vi.fn();
+
+    await userEvent.click(screen.getByRole("combobox"));
+    await userEvent.click(screen.getByText("Base"));
+
+    expect(screen.getByText("Switched to Base successfully!")).toBeInTheDocument();
+    expect(window.open).toHaveBeenCalledWith("https://app.morpho.org/base/earn", "_blank", "noopener,noreferrer");
+  });
+});
