+++ apps/lightweight/test/setup.ts
@@ -8,6 +8,10 @@ global.ResizeObserver = vi.fn().mockImplementation(() => ({
   disconnect: vi.fn(),
 }));
 
+window.HTMLElement.prototype.scrollIntoView = vi.fn();
+window.HTMLElement.prototype.hasPointerCapture = vi.fn();
+window.HTMLElement.prototype.releasePointerCapture = vi.fn();
+
 // runs a clean after each test case (e.g. clearing jsdom)
 afterEach(() => {
   cleanup();
